requirements.txt
setuptools==41.0.1 
sortedcontainers==2.1.0  
websocket-client==0.56.0 